+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version 3.1.2
 - `BacktraceData` allows to edit list of environment variables collected by `BacktraceAnnotations`
+- `SourceCode` object description for PII purpose
 
 ## Version 3.1.1
 - Prevent erroneously extending backtraceClient attributes with backtraceReport attributes.